refactor(search):added search index for availity sites PF-262
